migrate modular device profile integration tests
